src/app/requirements.txt
psycopg2-binary
databricks-sdk>=0.50.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
